#feat 18
add delete test book
add edit test book
add get test book
add list test all books by category
add list test book
add list test page book
add save test book

creation of BookControllerV1Test

+++ src/main/java/com/phoebus/library/book/service/DeleteBookServiceImpl.java
@@ -1,6 +1,7 @@
 package com.phoebus.library.book.service;
 
 import com.phoebus.library.book.BookRepository;
+import com.phoebus.library.exceptions.BookNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -12,6 +13,9 @@ public class DeleteBookServiceImpl implements DeleteBookService{
 
     @Override
     public void delete(Long id) {
+        if(!repository.existsById(id)){
+            throw new BookNotFoundException();
+        }
         repository.deleteById(id);
     }
 }
